Add test vectors for error inside blinded path
When a trampoline node inside a trampoline blinded path wants to fail an
incoming payment, it must use `update_fail_malformed_htlc` to avoid
leaking blinded path information. The introduction node will translate
that into a real `temporary_trampoline_failure` that it then encrypts to
the payer. We add a test vector that showcases this behavior.