Remove the CoC.md
The code of conduct as it has been established and how they are
enforced are in violation of BOLT contributors rights as set
by all the major jurisdictions from where are historical or active
individual contributors or organizational stakeholders to the
lightning bolt project.

For now, remove the Coc.md until there is legal clarity on
the role and extend of each contributor consent w.r.t any
hypothetical moderation rules.