Disallow blinded paths between intermediate trampoline nodes
There is no identified scenario where intermediate trampoline nodes use
a blinded path to reach the next trampoline node, since they know the
identity of that next node. We thus disallow this case by specifying
that the `current_path_key` can only be set in either the outer onion
or the trampoline onion (for the introduction node).