Introduce a `start_batch` message for `commit_sig`
Instead of repeating a `batch` TLV in each `commit_sig` message when
there are pending splice transactions, we prefix the batch of
`commit_sig` messages with a `start_batch` message. The following
`commit_sig` messages must be sent immediately afterwards, without
inserting any other unrelated message in that batch. This allows the
receiver to group the `commit_sig` messages when reading them from
the wire and process them together.

This mechanism is more general and may be reused for other scenarios
in the future.