Clarify Mandatory Field Length Requirements (#1243)
We previously allowed existing fields with unexpected lengths, and
required readers to ignore them. The goal was to provide a future
upgrade path that would reuse the same bech32 prefixes. However, we
don't have any plans for this, Bolt 12 introduced TLVs in invoices
which solves this problem entirely, and this was making the spec
unnecessarily complex. We thus remove those requirements.

We also add a note on low R signatures in test vectors.