Add requirements to avoid rebroadcasting spent channels
After a channel has been spent, we temporarily keep it in our graph in
case it was spent by a splice transaction: however, we must not
rebroadcast the corresponding `channel_announcement` to our peers, who
may otherwise think we're spamming them with invalid channels.

We don't rebroadcast `channel_update`s either in that case, unless they
have the `disable` bit set to 1, which indicates that the channel is
likely closed.