Clarify `tx_signatures` ordering
The specification for who sends `tx_signatures` first states that the
node that has the lowest satoshis contributed, "as defined by total
`tx_add_input` values" must sign first. Since the splice initiator is
sending `tx_add_input` for the current channel output, that means all
of the channel's previous capacity is attributed to the initiator,
instead of taking into account each node's pre-splice channel balance.

Since this detail is easy to miss, we add a note in the requirements
for `tx_signatures` in the splice section to highlight this subtlety.

Suggested by @wpaulino