Use `next_local_nonces` map wherever necessary
We must use the `next_local_nonces` TLV instead of a single nonce TLV
in the following messages to be future-proof with splicing:

- `channel_reestablish`
- `revoke_and_ack`

We must include one nonce for each commitment transaction that can be
published: when there is a pending splice, that means we need one nonce
for the current funding transaction, another nonce for the commitment
transaction that spends the splice transaction, and additional nonces
for each RBF attempt of that splice transaction (if any).

All of those commitment transactions use the same `per_commitment_point`
and are all revoked with the same secret: even though we send multiple
`commit_sig` messages while splices are pending, we send always send a
single `revoke_and_ack` message. This is why it must contain the map of
all next local nonces.