03: remove note that remote uses your `localpubkey` in their `to_remote`
As defined in the specification, our `localpubkey` is derived from
our `payment_basepoint` and the `per_commitment_point`, which means
that it changes for each commitment.

Now that we assume `option_static_remote`, we expect our counterparty
to always set their `to_remote` to use our static `payment_basepoint`
(not `localpubkey`) so that we can recover funds without knowledge of
the `per_commitment_point`.

This note contradicts that, making it seem like we're still using the
variable `localpubkey` in both the local and remote party's commitments,
not just the local one.